Updated the models to reflect proper ownership
Since it's my blog, I'm the only user that can post, but rather than making me a special user, I'll just not give any user that power. Posts will contain a reference to the post they actually own. This will keep the database lighter and also not require me to load in the documents each time I move the database.